chore(djf.io): upgrade Astro to v6 and refresh dependencies
Bumps astro 5.18 -> 6.3, @astrojs/mdx 4 -> 5, @astrojs/react 4 -> 5,
plus minor refreshes across the dep tree. Migrates the blog collection
off the legacy Content Collections v2 API (removed in Astro 6) to the
Content Layer API: glob loader, post.id in place of post.slug, and
render(post) in place of post.render(). Generated URLs unchanged.
Closes the upgrade-astro-versions project.
