+++ apps/djf.io/package.json
@@ -20,24 +20,24 @@
     "test:e2e:ui": "playwright test --ui"
   },
   "dependencies": {
-    "@astrojs/check": "^0.9.4",
-    "@astrojs/mdx": "^4.0.0",
-    "@astrojs/react": "^4.4.2",
-    "@astrojs/rss": "^4.0.15",
-    "@types/react": "^19.0.0",
-    "@types/react-dom": "^19.0.0",
-    "astro": "^5.18.1",
-    "react": "^19.0.0",
-    "react-dom": "^19.0.0",
-    "sharp": "^0.34.0",
-    "typescript": "^5.3.3"
+    "@astrojs/check": "^0.9.9",
+    "@astrojs/mdx": "^5.0.4",
+    "@astrojs/react": "^5.0.4",
+    "@astrojs/rss": "^4.0.18",
+    "@types/react": "^19.2.14",
+    "@types/react-dom": "^19.2.3",
+    "astro": "^6.3.1",
+    "react": "^19.2.6",
+    "react-dom": "^19.2.6",
+    "sharp": "^0.34.5",
+    "typescript": "^5.9.3"
   },
   "devDependencies": {
-    "@cspell/dict-typescript": "^3.2.0",
-    "@pandacss/dev": "1.7.2",
-    "@playwright/test": "^1.50.0",
-    "cspell": "^9.0.0",
-    "vitest": "^3.0.0"
+    "@cspell/dict-typescript": "^3.2.3",
+    "@pandacss/dev": "1.11.1",
+    "@playwright/test": "^1.60.0",
+    "cspell": "^9.4.0",
+    "vitest": "^3.2.4"
   },
   "engines": {
     "node": ">=24"
