+++ apps/djf.io/src/pages/blog/[...slug].astro
@@ -1,17 +1,17 @@
 ---
-import {getCollection} from 'astro:content'
+import {getCollection, render} from 'astro:content'
 import BlogPost from '../../layouts/BlogPost.astro'
 
 export async function getStaticPaths() {
   const posts = await getCollection('blog')
   return posts.map((post) => ({
-    params: {slug: post.slug},
+    params: {slug: post.id},
     props: {post},
   }))
 }
 
 const {post} = Astro.props
-const {Content} = await post.render()
+const {Content} = await render(post)
 ---
 
 <BlogPost post={post}>
