+++ apps/djf.io/src/pages/blog/tags/[tag].astro
@@ -74,7 +74,7 @@ const {tag, posts} = Astro.props
             })}
           >
             <a
-              href={`/blog/${post.slug}`}
+              href={`/blog/${post.id}`}
               class={css({textDecoration: 'none', display: 'block'})}
             >
               <h2
