+++ docs/changelog/2026-05.md
@@ -2,6 +2,10 @@
 
 ## 2026-05-13
 
+### chore(djf.io): upgrade Astro to v6 and refresh dependencies
+
+Bumped `astro` `^5.18.1` → `^6.3.1`, `@astrojs/mdx` `^4` → `^5.0.4`, `@astrojs/react` `^4.4.2` → `^5.0.4`, `@astrojs/rss` `^4.0.15` → `^4.0.18`, `@astrojs/check` `^0.9.4` → `^0.9.9`, plus minor bumps to `react`, `react-dom`, `@types/react*`, `sharp`, `typescript`, `@playwright/test`, `@cspell/dict-typescript`, `cspell`, `vitest`, and `@pandacss/dev`. Migrated the blog collection off the legacy Content Collections v2 API (removed in Astro 6) to the Content Layer API: config moved from `src/content/config.ts` to `src/content.config.ts`, schema switched to a `glob({pattern: '**/*.md', base: './src/content/blog'})` loader, and the schema function extracted as `blogSchema` for direct test access. Updated consumers to read `post.id` instead of `post.slug` (`/blog/[...slug].astro`, `rss.xml.ts`, `blog/index.astro`, `blog/tags/[tag].astro`) and switched `post.render()` to the imported `render(post)` from `astro:content`. Moved `z` imports from the deprecated `astro:content` re-export to `astro/zod`. Generated URLs are unchanged (`/blog/2025-12-07-on-running/` etc.). Closed out `docs/projects/upgrade-astro-versions/`.
+
 ### docs(repo): file pending human-intervention issues and remove draft file
 
 Filed the four drafted human-intervention issues against `DavidJFelix/DavidJFelix` once the GitHub MCP server was reachable: install Warp CLI (#93), create Sentry org+projects (#94), create PostHog org+projects (#95), davidjfelix.com DNS cutover (#96). Deleted `docs/pending-issues.md` — issues are the durable record.
