--- docs/projects/upgrade-astro-versions/2026-04-30-progress.md
@@ -1,14 +1,0 @@
-# 2026-04-30 — Project kickoff
-
-Created project. Scope is narrower than the name suggests: only djf.io needs upgrading.
-
-## Scope clarified
-
-- djf.io is currently on `astro@^5.18.1` ([apps/djf.io/package.json](../../../apps/djf.io/package.json)).
-- Target: Astro 6.
-- Bump official `@astrojs/*` integrations alongside.
-
-## Next
-
-- Read the Astro 5 → 6 migration guide.
-- Bump versions and work through any breakage.
